+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 Scramjet 4 brings a stable interface, multi-threading, CSV parsing, asynchronous data augmentation and on...
 
-## Scramjet 4.18.2 - do and from
+## Scramjet 4.18.4 - do and from
 
+* Fix async iterator and generator behavior in `DataStream.from`.
 * Fix error handling so it doesn't need to return an array to work.
 * Errors now carry `cause` field carrying the original error object.
 * Add new options to pass to the `from` method - now anything sensible may be used to create a stream.